+++ GIT_UART_OXIMETER_filter/Filters.h
@@ -30,6 +30,6 @@ void Accumulator_Init_values(uint8_t);
 
 //float Set_EMA_Alpha();
 
-void linear_Regression(int32_t x[],uint16_t n,float *a,float *b,float *r);
-void Detrend(int32_t y[],uint16_t n,float *a,float *b);
-int Linear_Regression1(int32_t y[],uint16_t n,float *a,float *b,float *r);
+void linear_Regression(int32_t x[],uint16_t n,double *a,double *b,double *r);
+void Detrend(int32_t y[],uint16_t n,double *a,double *b);
+int Linear_Regression1(int32_t y[],uint16_t n,double *a,double *b,double *r);
